Extract Analyzer (but not Generator) into analysis.py
